Don't pass the group information to extract_features.py
I think it would be nice that we would strip the groups away after doing
the split of training/test data based on blocks. This way we would only need
to do it once, and we could directly use the features returned by these
functions in the sklearn classification fitting phase "fit()".